DPSWarr - Took out expertise rounding, updated dual-wield miss and glance
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C32471

+++ Rawr.DPSWarr/CalculationsDPSWarr.cs
@@ -176,8 +176,8 @@ namespace Rawr.DPSWarr
             calculatedStats.AddRoundedDisplayValue(DisplayValue.ArmorPenetration, stats.ArmorPenetrationRating);
             calculatedStats.AddRoundedDisplayValue(DisplayValue.HitPercent, stats.HitRating);
             calculatedStats.AddRoundedDisplayValue(DisplayValue.CritPercent, stats.CritRating);
-            calculatedStats.AddRoundedDisplayValue(DisplayValue.MhExpertise, (int)combatFactors.MhExpertise);
-            calculatedStats.AddRoundedDisplayValue(DisplayValue.OhExpertise, (int)combatFactors.OhExpertise);
+            calculatedStats.AddRoundedDisplayValue(DisplayValue.MhExpertise, combatFactors.MhExpertise);
+            calculatedStats.AddRoundedDisplayValue(DisplayValue.OhExpertise, combatFactors.OhExpertise);
             calculatedStats.AddRoundedDisplayValue(DisplayValue.Armor, (int)stats.Armor);
 
             calculatedStats.AddRoundedDisplayValue(DisplayValue.buffedMhCrit, combatFactors.MhCrit*100);
